2020-08-17 MSS Code Factory CFAll 2.13.11195-202008172000 All manufactured Java code is now deposited where it belongs
All of the manufactured 2.13 Java code is now in the properly named locations, directorories,
and directory layouts to support Maven builds.

org.msscf.msscf.v2_13.cflib.CFLib, org.msscf.msscf.v2_13.cflib.CFLib.JavaFX, and
org.msscf.msscf.v2_13.cfcore have been built and installed to my local system
using Maven.

I am ready to start pruning code from the 2.12 knowledge base and 2.13 projects now.

For starters, only the Ram and PgSql persistence implementations will carry forward;
the other in the 2.12 knowledge base are being moved to a private github repository,
and won't be used any more.  That includes the C++ code rules; I'm not happy with the
performance, so I don't want anyone using it and I don't have time to deal with it right
now.  I won't be deleting the C++ code in the cfsec, cfint, cfbam, and msscf 2.13 projects,
though - there is a lot of custom code in there I want to hang on to; I just won't be
building or maintaining it for the foreseeeable future.

I won't be removing any more 2.12 models - I'm content that I've pruned the project set back
to the original group I was working with nearly a decade ago, before I got
side-tracking with work-specific models in additiion to those essentials. The only other
one to survive is CFAcc as a private project; I have ideas for that that go beyond actual
accounting, and apply specifically to some needs I have for my code in the future.

The only thing other than the manufactured code in the Java directories are the old build.xml
files, the Package scripts, and the custom code modules in their old Ant locations.  I'll
deal with those after I produce my POMs for the code layers I'm keeping.